% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxEval.R
\docType{data}
\name{ACC}
\alias{ACC}
\title{ACC values included with toxEval.}
\description{
Data frame containing concentrations for the ToxCast activity concentration at cutoff (ACC). 
Downloaded on October 2015 from ToxCast. The data were
combined from files in the "INVITRODB_V2_LEVEL5" folder. 
At the time of toxEval package release, this information was found:
\url{https://www.epa.gov/chemical-research/toxicity-forecaster-toxcasttm-data}
in the "ToxCast & Tox21 Data Spreadsheet" data set.
}
\details{
Columns include the Chemical Abstract Service Registry Number (CAS), the chemical name (chnm), data quality flags (flags)as the first three columns. The remaining columns are ACC values for individual endpoints (column names) for each chemical represented. The data has been provided in a "wide" format; however, the \code{get_ACC} function provides a method to transform the data into a "long" format. ACC values are the reported modl_ga (winning model) and units are in log micro-Molarity (log $\mu$M).
}
\examples{
ACCColumnNames <- names(ACC)
}
\keyword{datasets}
