% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/toxEval.R
\docType{data}
\name{assayEndpoints}
\alias{assayEndpoints}
\title{Assay Selection data from ToxCast}
\description{
Downloaded on March 20, 2015 from ToxCast dashboard
}
\examples{
assayEndpoints <- assayEndpoints
head(assayEndpoints)
}
\keyword{datasets}

