% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chemSumm.R
\name{chemSumm}
\alias{chemSumm}
\title{Create chemical summary}
\usage{
chemSumm(chemicalSummary)
}
\arguments{
\item{chemicalSummary}{data frame returned from \code{chemSummBasic}}
}
\description{
Create chemical summary
}
\examples{
library(dplyr)
wData <- wData
pCodeInfo <- pCodeInfo
endPoint <- endPointToxCreate(pCodeInfo)
chemicalSummary <- chemSummBasic(wData,pCodeInfo,endPoint)
chemSum1 <- chemSumm(chemicalSummary)
}

