% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_endPoint_info.R
\name{clean_endPoint_info}
\alias{clean_endPoint_info}
\title{clean_endPoint_info}
\usage{
clean_endPoint_info(end_point_info)
}
\arguments{
\item{end_point_info}{Data frame Endpoint information from ToxCast.}
}
\value{
The returned data frame is based on end_point_info, but with some endPoints
filtered out and some additional categories in intended_target_family and
intended_target_family_sub. The names in intended_target_family
are revised to look more appealing in graphs and tables.
}
\description{
As of ToxCast 4.1, this function only helps clean up abbrieviations 
found in the end_point_info data frame.
}
\examples{
end_point_info <- end_point_info
nrow(end_point_info)
cleaned_ep <- clean_endPoint_info(end_point_info)
nrow(cleaned_ep)
}
