% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxEval.R
\docType{data}
\name{Constants}
\alias{AC10}
\alias{AC50gain}
\alias{AC50loss}
\alias{ACC}
\alias{Constants}
\title{Constants included with toxEval. Units are log micro-Molarity (log uM).}
\description{
AC50gain downloaded on October 2015 from ToxCast dashboard. AC50gain values 
are either the reported modl_ga (winning model) or 10% of modl_ga if the AC50gain
value is lower than the lowest measured concentration. Also,
}
\details{
\itemize{
 \item{AC50gain}{AC50 gain endpoints}
 \item{AC50loss}{AC50 loss endpoints}
 \item{AC10}{AC10 endpoints}
}
}
\examples{
AC50GainColumns <- names(AC50gain)
AC50LossColumns <- names(AC50loss)
AC10Columns <- names(AC10)
}
\keyword{datasets}

