% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_toxEval.R
\name{create_toxEval}
\alias{create_toxEval}
\title{create_toxEval}
\usage{
create_toxEval(excel_file_path)
}
\arguments{
\item{excel_file_path}{Path to Excel file that contains at least 3 tabs: Data, Chemicals, and Sites, 
and could optionally contain Exclude and Benchmark}
}
\description{
create_toxEval
}
\examples{
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"
excel_file_path <- file.path(path_to_tox, file_name)
tox_list <- create_toxEval(excel_file_path)
}
