% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_endpoint_hits.R
\name{createLink}
\alias{createLink}
\title{createLink}
\usage{
createLink(cas, endpoint, hits)
}
\arguments{
\item{cas}{character}

\item{endpoint}{character}

\item{hits}{character}
}
\description{
Create links
}
\keyword{internal}
