% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint_hits.R
\name{endpoint_hits_DT}
\alias{endpoint_hits_DT}
\alias{endpoint_hits}
\title{Rank endpoints by category}
\usage{
endpoint_hits_DT(chemicalSummary, category = "Biological",
  mean_logic = FALSE, hit_threshold = 0.1)

endpoint_hits(chemicalSummary, category = "Biological", mean_logic = FALSE,
  hit_threshold = 0.1)
}
\arguments{
\item{chemicalSummary}{data frame from \code{get_chemical_summary}}

\item{category}{either "Biological", "Chemical Class", or "Chemical"}

\item{mean_logic}{logical \code{TRUE} is mean, \code{FALSE} is maximum}

\item{hit_threshold}{numeric threshold defining a "hit"}
}
\description{
These functions create a table with one row per endPoint, and one 
column per category("Biological", "Chemical", or "Chemical Class"). 
The values in the table are number of sites with hits in that endpoint/category 
combination, based on a user-specified threshold (defined by hit_threshold).
}
\details{
The tables show slightly different results for a single site. Instead of the 
number of sites with hits above a threshold, it is now the number of samples with hits.
}
\examples{
# This is the example workflow:
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"

full_path <- file.path(path_to_tox, file_name)
\dontrun{
tox_list <- create_toxEval(full_path)

ACClong <- get_ACC(tox_list$chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)
chemicalSummary <- get_chemical_summary(tox_list, ACClong, filtered_ep)

hits_df <- endpoint_hits(chemicalSummary, category = "Biological")                        
endpoint_hits_DT(chemicalSummary, category = "Biological")
endpoint_hits_DT(chemicalSummary, category = "Chemical Class")
endpoint_hits_DT(chemicalSummary, category = "Chemical")
}
}
