% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxEval.R
\docType{data}
\name{endPointInfo}
\alias{endPointInfo}
\alias{flagDF}
\title{Endpoint information from ToxCast}
\description{
Downloaded on October 2015 from ToxCast dashboard
}
\examples{
endPointInfo <- endPointInfo
head(endPointInfo[,1:5])
flagDF <- flagDF
head(flagDF)
}
\keyword{datasets}

