% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/endPointSumm.R
\name{endPointSumm}
\alias{endPointSumm}
\title{Create chemical summary}
\usage{
endPointSumm(chemicalSummary, chemicalSummary.Key = "endPoint",
  chemicalSummary.site = "site", endPointInfo,
  endPointInfo.Key = "assay_component_endpoint_name", EAR.key = "EAR")
}
\arguments{
\item{chemicalSummary}{data frame returned from \code{chemSummBasic}}

\item{chemicalSummary.Key}{character column name for grouping by (default = "endPoint")}

\item{chemicalSummary.site}{character column name for group by site}

\item{endPointInfo}{dataframe}

\item{endPointInfo.Key}{column name of endpoint}

\item{EAR.key}{column name in chemicalSummary of EAR}
}
\description{
Create chemical summary
}
\examples{
\dontrun{
wData <- wData
pCodeInfo <- pCodeInfo
endPoint <- endPointToxCreate(pCodeInfo)
chemicalSummary <- chemSummBasic(wData,pCodeInfo,endPoint)
endPointInfo <- endPointInfo
epSumm <- endPointSumm(chemicalSummary, endPointInfo=endPointInfo)
}
}

