% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemSumm.R
\name{endPointToxCreate}
\alias{endPointToxCreate}
\title{Create endpoint dataframe}
\usage{
endPointToxCreate(pCodeInfoDF, casrn_pCode = "casrn",
  units_pCode = "parameter_units", mlWt_pCode = "mlWt")
}
\arguments{
\item{pCodeInfoDF}{data frame}

\item{casrn_pCode}{character name of casrn column in pCodeInfo}

\item{units_pCode}{character name of units column in pCodeInfo}

\item{mlWt_pCode}{character name of molecular weight column in pCodeInfo}
}
\value{
endPoint data frame
}
\description{
Create endpoint dataframe
}
\examples{
pCodeInfo <- pCodeInfo
endPoint <- endPointToxCreate(pCodeInfo)
}

