% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxEval.R
\docType{data}
\name{ex_chemSum}
\alias{ex_chemSum}
\title{Example chemicalSummary}
\description{
Example data representing data from the sample Excel file found in
\code{system.file("extdata", package="toxEval")}, OWC_data_fromSup.xlsx
}
\examples{
head(ex_chemSum)
}
\keyword{data}
\keyword{quality}
\keyword{water}
