% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chemical_summary.R
\name{exclude_points}
\alias{exclude_points}
\title{Exclude endPoint/Chem combos}
\usage{
exclude_points(chemicalSummary, exclusion)
}
\arguments{
\item{chemicalSummary}{data frame from \code{graph_chem_data}}

\item{exclusion}{data frame with columns "CAS" and "endPoint"}
}
\description{
Using a dataframe "exclusion", filter out all chemical/endpoint combos
}
\examples{
library(readxl)
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"
full_path <- file.path(path_to_tox, file_name)

chem_data <- read_excel(full_path, sheet = "Data")
chem_info <- read_excel(full_path, sheet = "Chemicals") 
chem_site <- read_excel(full_path, sheet = "Sites")
exclusion <- read_excel(full_path, sheet = "Exclude")
ACClong <- get_ACC(chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)

chemicalSummary <- get_chemical_summary(ACClong,
                                        filtered_ep,
                                        chem.data = chem_data, 
                                        chem.site = chem_site, 
                                        chem.info = chem_info)
chemicalSummary <- exclude_points(chemicalSummary, exclusion)
}
