% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_endpoints.R
\name{explore_endpoints}
\alias{explore_endpoints}
\title{Explore data in the Shiny Application}
\usage{
explore_endpoints(browse = TRUE)
}
\arguments{
\item{browse}{use browser for map rendering}
}
\description{
Open an interactive app in a browser. See the vignette 'User Guide'
for more details.
}
\examples{
\dontrun{
explore_endpoints()
}
}
