% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group_boxplots.R
\name{fancyNumbers2}
\alias{fancyNumbers2}
\title{fancyNumbers2}
\usage{
fancyNumbers2(n)
}
\arguments{
\item{n}{vectore}
}
\description{
Just another fancy ggplot2 axis labeler.
}
\keyword{internal}
