% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_endPoint_info.R
\name{filter_flags}
\alias{filter_flags}
\title{filter_flags}
\usage{
filter_flags(ep, flagsShort = c("Borderline", "OnlyHighest", "GainAC50",
  "Biochemical"))
}
\arguments{
\item{ep}{data frame Endpoint information from ToxCast}

\item{flagsShort}{vector of flags TO REMOVE}
}
\description{
Clean up the endPointInfo table from toxCast. Filtering and cleaning based on ES&T (cite Dan/Brett's paper)
}
\examples{
endPointInfo <- endPointInfo
cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)

}
