% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_endPoint_info.R, R/get_ACC.R
\name{filter_groups}
\alias{filter_groups}
\alias{filter_groups}
\title{filter_groups}
\usage{
filter_groups(ep, groupCol = "intended_target_family", assays = c("ATG",
  "NVS", "OT", "TOX21", "CEETOX", "APR", "CLD", "TANGUAY", "NHEERL_PADILLA",
  "NCCT_SIMMONS", "ACEA"), remove_group = "Background Measurement")

filter_groups(ep, groupCol = "intended_target_family", assays = c("ATG",
  "NVS", "OT", "TOX21", "CEETOX", "APR", "CLD", "TANGUAY", "NHEERL_PADILLA",
  "NCCT_SIMMONS", "ACEA"), remove_group = "Background Measurement")
}
\arguments{
\item{ep}{data frame Endpoint information from ToxCast}

\item{groupCol}{character name of column to use as a group catetory}

\item{assays}{vector of assays to use. Possible values are "ATG","NVS","OT","TOX21","CEETOX", "APR", "BSK",
"CLD","TANGUAY","NHEERL_PADILLA","NCCT_SIMMONS","ACEA"}

\item{remove_groups}{vector of groups to remove}

\item{ep}{data frame Endpoint information from ToxCast}

\item{flagsShort}{vector of flags TO REMOVE}
}
\description{
Clean up the endPointInfo table from toxCast. Filtering and cleaning based on ES&T (cite Dan/Brett's paper)

Clean up the endPointInfo table from toxCast. Filtering and cleaning based on ES&T (cite Dan/Brett's paper)
}
\examples{
endPointInfo <- endPointInfo
cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)
endPointInfo <- endPointInfo
cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)
}
