% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxEval.R
\docType{data}
\name{flags}
\alias{flags}
\title{ToxCast Chemical Information}
\value{
data frame
}
\description{
See \code{vignette("Setting up toxEval package data", package = "toxEval")}
for more information on how the data was aggregated.
}
\examples{
head(flags)
}
\keyword{datasets}
