% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ACC.R
\name{get_ACC}
\alias{get_ACC}
\title{get_ACC}
\usage{
get_ACC(CAS)
}
\arguments{
\item{CAS}{vector of CAS}
}
\description{
Get ACC values for vector of CAS's
}
\examples{
CAS <- c("121-00-6","136-85-6","80-05-7","84-65-1","5436-43-1","126-73-8")
ACClong <- get_ACC(CAS)
}
