% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chemical_boxplots.R
\name{graph_chem_data}
\alias{graph_chem_data}
\title{graph_chem_data}
\usage{
graph_chem_data(chemicalSummary, manual_remove = NULL, mean_logic = FALSE)
}
\arguments{
\item{chemicalSummary}{data frame from \code{graph_chem_data}}

\item{manual_remove}{vector of categories to remove}

\item{mean_logic}{logical \code{TRUE} is mean, \code{FALSE} is maximum}
}
\description{
Get chemical data summarized for plots
}
\examples{
# This is the example workflow:
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"

full_path <- file.path(path_to_tox, file_name)

tox_list <- create_toxEval(full_path)
\dontrun{
ACClong <- get_ACC(tox_list$chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)
chemicalSummary <- get_chemical_summary(ACClong, filtered_ep,
                                        tox_list)
}
# The example workflow takes a bit of time to load and compute, 
# so an example chemicalSummary is included pre-calculated in the package. 
chemicalSummary <- ex_chemSum #loading example data
graphData <- graph_chem_data(chemicalSummary)
}
