% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group_boxplots.R
\name{graphData}
\alias{graphData}
\title{graphData}
\usage{
graphData(chemicalSummary, category = "Biological", manual_remove = NULL,
  mean_logic = FALSE)
}
\arguments{
\item{chemicalSummary}{data frame}

\item{category}{character}

\item{manual_remove}{vector}

\item{mean_logic}{logical}
}
\description{
Summarize data for most graphs/tables
}
\keyword{internal}
