% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chemical_summary.R
\name{orderChem}
\alias{orderChem}
\title{orderChem}
\usage{
orderChem(graphData, orderClass_df)
}
\arguments{
\item{graphData}{data frame}

\item{orderClass_df}{data frame}
}
\description{
orderChem
}
