% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group_boxplots.R
\name{plot_group_boxplots}
\alias{plot_group_boxplots}
\title{plot_group_boxplots}
\usage{
plot_group_boxplots(chemicalSummary, filtered_ep, category = "Biological",
  manual_remove = NULL)
}
\arguments{
\item{chemicalSummary}{data frame from \code{get_chemical_summary}}

\item{filtered_ep}{data frame from \code{filter_groups}}

\item{category}{either "Biological" or "Chemical Class"}

\item{manual_remove}{vector of categories to remove}
}
\description{
Plot boxplot of groups
}
\examples{
CAS <- c("121-00-6","136-85-6","80-05-7","84-65-1","5436-43-1","126-73-8")
ACClong <- get_ACC(CAS)
ACClong <- remove_flags(ACClong)
cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)

}
