% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tox_stacks.R
\name{plot_tox_stacks}
\alias{plot_tox_stacks}
\title{Plot stacked bar charts}
\usage{
plot_tox_stacks(chemicalSummary, chem_site, category = "Biological",
  mean_logic = FALSE, manual_remove = NULL, include_legend = TRUE,
  font_size = NA, title = NA)
}
\arguments{
\item{chemicalSummary}{data frame from \code{get_chemical_summary}}

\item{chem_site}{data frame with at least columns SiteID, site_grouping,  and Short Name}

\item{category}{either "Biological", "Chemical Class", or "Chemical"}

\item{mean_logic}{logical \code{TRUE} is mean, \code{FALSE} is maximum}

\item{manual_remove}{vector of categories to remove}

\item{include_legend}{logical to include legend or not}

\item{font_size}{numeric to adjust the axis font size}

\item{title}{character title for plot.}
}
\description{
This function creates a set of stacked bar charts based on the original 
input data modified by the processing steps above, and the choice of several 
input options. The result includes stacked bar charts of the maximum or mean 
value per site based on the mean_logic argument. This ensures that each site 
is represented equally regardless of how many samples are available per site. 
Stacked bar charts are generated by the chosen category. Categories include 
"Biological", "Chemical Class", or "Chemical". "Biological" refers to the chosen 
ToxCast annotation as defined in the groupCol argument of the \code{filter_groups}
function. "Chemical Class" refers to the groupings of chemicals as defined in the 
"Class" column of the "Chemicals" tab of the input file. Choosing "Chemical Class" 
will generate a separate stacked bar chart for each unique class. Choosing "Chemical" 
will generate a separate stacked bar chart for each individual chemical in the data 
set.
}
\details{
The graph shows a slightly different result for a single site. The single-site plots 
now shows each individual sample. The EAR values are no longer the mean or max at 
the site, but rather the sums of the individual EAR values.
}
\examples{
# This is the example workflow:
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"

full_path <- file.path(path_to_tox, file_name)

tox_list <- create_toxEval(full_path)
\dontrun{
ACClong <- get_ACC(tox_list$chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)
chemicalSummary <- get_chemical_summary(tox_list, ACClong, filtered_ep)
                                       
plot_tox_stacks(chemicalSummary, tox_list$chem_site, "Biological")   
plot_tox_stacks(chemicalSummary, tox_list$chem_site, "Chemical Class")
plot_tox_stacks(chemicalSummary, tox_list$chem_site, "Chemical", include_legend = FALSE) 

# To turn off clipping:
class_plot <- plot_tox_stacks(chemicalSummary, "Chemical Class")
gb <- ggplot2::ggplot_build(class_plot)
gt <- ggplot2::ggplot_gtable(gb)

gt$layout$clip[gt$layout$name=="panel-1-1"] <- "off"

grid::grid.draw(gt) 
}
}
