% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_sites.R
\name{rank_sites_DT}
\alias{rank_sites_DT}
\alias{rank_sites}
\title{Rank sites by EAR}
\usage{
rank_sites_DT(chemicalSummary, category = "Biological", mean_logic = FALSE,
  sum_logic = TRUE, hit_threshold = 0.1)

rank_sites(chemicalSummary, category, hit_threshold = 0.1,
  mean_logic = FALSE, sum_logic = TRUE)
}
\arguments{
\item{chemicalSummary}{data frame from \code{get_chemical_summary}}

\item{category}{either "Biological", "Chemical Class", or "Chemical"}

\item{mean_logic}{logical.  TRUE takes the mean sample of each site,
FALSE takes the maximum sample of each site.}

\item{sum_logic}{logical. TRUE sums the EARs in a specified grouping,
FALSE does not. FALSE may be better for traditional benchmarks as
opposed to ToxCast benchmarks.}

\item{hit_threshold}{numeric threshold defining a "hit"}
}
\description{
These functions create a table (either data frame or DT) with one row per site. 
There are columns of the max or mean EAR (depending on the 'mean_logic' argument) 
for each category ("Chemical Class", "Chemical", or "Biological"). Additionally, 
there are columns of the frequency of the max or mean EAR being above a user 
specified 'hit_threshold'.
}
\details{
The tables show slightly different results for a single site. Instead of multiple 
columns for category, there is now 1 row per category (since the site is known).
}
\examples{
# This is the example workflow:
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"

full_path <- file.path(path_to_tox, file_name)

tox_list <- create_toxEval(full_path)
\dontrun{
ACClong <- get_ACC(tox_list$chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)
chemicalSummary <- get_chemical_summary(tox_list, ACClong, filtered_ep)

stats_df <- rank_sites(chemicalSummary, "Biological")

rank_sites_DT(chemicalSummary, category = "Biological")
rank_sites_DT(chemicalSummary, category = "Chemical Class")
rank_sites_DT(chemicalSummary, category = "Chemical")
}
}
