% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chemical_summary.R
\name{remove_flags}
\alias{remove_flags}
\title{remove_flags}
\usage{
remove_flags(ACClong, flagsShort = c("Borderline", "OnlyHighest", "GainAC50",
  "Biochemical"))
}
\arguments{
\item{ACClong}{data frame with columns: casn, chnm, endPoint, ACC_value}

\item{flagsShort}{vector of flags to TAKE OUT. Possible values are 
"Borderline", "OnlyHighest", "OneAbove","Noisy", "HitCall", "GainAC50", "Biochemical"}
}
\description{
Remove endpoints with specific flags from data
}
\examples{
CAS <- c("121-00-6","136-85-6","80-05-7","84-65-1","5436-43-1","126-73-8")
ACClong <- get_ACC(CAS)
ACClong <- remove_flags(ACClong)
}
