% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siteSumm.R
\name{siteSumm}
\alias{siteSumm}
\title{Create site summary}
\usage{
siteSumm(chemicalSummary, newSiteKey)
}
\arguments{
\item{chemicalSummary}{data frame returned from \code{chemSummBasic}}

\item{newSiteKey}{named vector}
}
\description{
Create site summary
}
\examples{
wData <- wData
pCodeInfo <- pCodeInfo
packagePath <- system.file("extdata", package="toxEval")
filePath <- file.path(packagePath, "stationINFO.RData")
load(file=filePath)
newSiteKey <- setNames(stationINFO$shortName, stationINFO$fullSiteID)
endPoint <- endPointToxCreate(pCodeInfo)
chemicalSummary <- chemSummBasic(wData,pCodeInfo,endPoint)
siteSummary <- siteSumm(chemicalSummary,newSiteKey)
}

