% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_tox_rank.R
\name{statsOfColumns}
\alias{statsOfColumns}
\title{statsOfColumns}
\usage{
statsOfColumns(chemicalSummary, category, hit_threshold, mean_logic)
}
\arguments{
\item{chemicalSummary}{data frame}

\item{category}{character}

\item{hit_threshold}{numeric}

\item{mean_logic}{logical}
}
\description{
Summarize data for most graphs/tables
}
\keyword{internal}
