% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_tox_sum.R
\name{statsOfGroup}
\alias{statsOfGroup}
\title{statsOfGroup}
\usage{
statsOfGroup(chemicalSummary, category, hit_threshold)
}
\arguments{
\item{chemicalSummary}{data frame}

\item{category}{character}

\item{hit_threshold}{numeric}
}
\description{
Summarize data for most graphs/tables
}
\keyword{internal}
