% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_toxEval.R
\name{summary.toxEval}
\alias{summary.toxEval}
\title{summary of tox_list}
\usage{
\method{summary}{toxEval}(x, ...)
}
\arguments{
\item{x}{toxEval object with "chem_info" data frame included.}

\item{\dots}{additional parameters}
}
\description{
summary of tox_list
}
\examples{
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"
excel_file_path <- file.path(path_to_tox, file_name)
tox_list <- create_toxEval(excel_file_path)
summary(tox_list) 
}
