% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_tox_sum.R
\name{table_tox_sum}
\alias{table_tox_sum}
\title{table_tox_sum}
\usage{
table_tox_sum(chemicalSummary, category = "Biological", mean_logic = FALSE,
  hit_threshold = 0.1)
}
\arguments{
\item{chemicalSummary}{data frame from \code{get_chemical_summary}}

\item{category}{either "Biological", "Chemical Class", or "Chemical"}

\item{mean_logic}{logical \code{TRUE} is mean, \code{FALSE} is maximum}

\item{hit_threshold}{numeric threshold defining a "hit"}
}
\description{
Table of sums
}
\examples{
# This is the example workflow:
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"

full_path <- file.path(path_to_tox, file_name)

tox_list <- create_toxEval(full_path)
\dontrun{
ACClong <- get_ACC(tox_list$chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)
chemicalSummary <- get_chemical_summary(tox_list, ACClong, filtered_ep)

table_tox_sum(chemicalSummary, category = "Biological")
table_tox_sum(chemicalSummary, category = "Chemical Class")
table_tox_sum(chemicalSummary, category = "Chemical")
}
}
