% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/totalSamples.R
\name{totalSamples}
\alias{totalSamples}
\title{Create total samples info}
\usage{
totalSamples(chemicalSummary, chemicalSummary.Key = "endPoint",
  chemicalSummary.site = "site")
}
\arguments{
\item{chemicalSummary}{data frame returned from \code{chemSummBasic}}

\item{chemicalSummary.Key}{character column name for grouping by (default = "endPoint")}

\item{chemicalSummary.site}{character column name for group by site}
}
\description{
Create total samples
}
\examples{
wData <- wData
pCodeInfo <- pCodeInfo
endPoint <- endPointToxCreate(pCodeInfo)
chemicalSummary <- chemSummBasic(wData,pCodeInfo,endPoint)
endPointInfo <- endPointInfo
epSumm <- totalSamples(chemicalSummary)
}

