% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxEval.R
\docType{data}
\name{tox_chemicals}
\alias{tox_chemicals}
\title{ToxCast Chemical Information}
\value{
data frame with columns:
"Substance_Name","Substance_CASRN",
"Structure_MolWt"
}
\description{
Downloaded from the CompTox database on October 2022. :
\url{https://comptox.epa.gov/dashboard/}.
}
\examples{
head(tox_chemicals)
}
\keyword{datasets}
