% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxEval.R
\docType{data}
\name{ACC}
\alias{ACC}
\title{ACC values included with toxEval.}
\description{
Downloaded on October 2015 from ToxCast. The data were
combined from files in the "INVITRODB_V2_LEVEL5" folder. 
At the time of toxEval package release, this information was found:
\url{https://www.epa.gov/chemical-research/toxicity-forecaster-toxcasttm-data}
in the "ToxCast & Tox21 Data Spreadsheet" data set. 
ACC values are the reported ACC_value (winning model) and units are 
log micro-Molarity (log \eqn{\mu}M).
}
\examples{
ACCColumnNames <- names(ACC)
}
\keyword{datasets}
