% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_endPoint_info.R
\name{clean_endPoint_info}
\alias{clean_endPoint_info}
\title{clean_endPoint_info}
\usage{
clean_endPoint_info(endPointInfo)
}
\arguments{
\item{endPointInfo}{data frame Endpoint information from ToxCast}
}
\description{
Clean up the endPointInfo table from toxCast. Filtering and cleaning based on ES&T (cite Dan/Brett's paper)
}
\examples{
endPointInfo <- endPointInfo
cleaned_ep <- clean_endPoint_info(endPointInfo)
}
