% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxEval.R
\docType{data}
\name{Constants}
\alias{Constants}
\alias{ACC}
\title{ACC values included with toxEval.}
\details{
AC50gain downloaded on October 2015 from ToxCast dashboard. AC50gain values 
are either the reported modl_ga (winning model) or 10% of modl_ga if the AC50gain
value is lower than the lowest measured concentration. Also,

Units are log micro-Molarity (log uM).

\itemize{
 \item{ACC}{ACC endpoints}
}
}
\examples{
ACCColumnNames <- names(ACC)
}
\keyword{datasets}
