% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_tox_endpoint.R
\name{endpoint_table}
\alias{endpoint_table}
\title{endpoint_table}
\usage{
endpoint_table(chemicalSummary, category, mean_logic = FALSE,
  hit_threshold = 0.1)
}
\arguments{
\item{chemicalSummary}{data frame from \code{get_chemical_summary}}

\item{category}{either "Biological", "Chemical Class", or "Chemical"}

\item{mean_logic}{logical \code{TRUE} is mean, \code{FALSE} is maximum}

\item{hit_threshold}{numeric threshold defining a "hit"}
}
\description{
Table of ranks
}
\examples{
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"
full_path <- file.path(path_to_tox, file_name)

\dontrun{
tox_list <- create_toxEval(full_path)
ACClong <- get_ACC(tox_list$chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)
chemicalSummary <- get_chemical_summary(tox_list, ACClong, filtered_ep)
                                          
bio_table <- endpoint_table(chemicalSummary, category = "Biological")
class_table <- endpoint_table(chemicalSummary, category = "Chemical Class")
chem_table <- endpoint_table(chemicalSummary, category = "Chemical")
}
}
