% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/toxEval.R
\docType{data}
\name{endPointInfo}
\alias{endPointInfo}
\title{Endpoint information from ToxCast}
\description{
Downloaded on March 20, 2015 from ToxCast dashboard
}
\examples{
endPointInfo <- endPointInfo
head(endPointInfo[,1:5])
}
\keyword{datasets}

