% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxEval.R
\docType{data}
\name{endPointInfo}
\alias{endPointInfo}
\title{Endpoint information from ToxCast}
\description{
Downloaded on October 2015 from ToxCast. The file name of the
raw data was "Assay_Summary_151020.csv" from the zip file 
"Assay_Information_Oct_2015.zip". At the time
of toxEval package release, this information was found:
\url{https://www.epa.gov/chemical-research/toxicity-forecaster-toxcasttm-data}
in the section marked "Download Assay Information", in the 
ToxCast & Tox21 high-throughput assay information data set.
}
\examples{
endPointInfo <- endPointInfo
head(endPointInfo[,1:5])
}
\keyword{datasets}
