% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chemSumm.R
\name{endPointToxCreate}
\alias{endPointToxCreate}
\title{Create chemical summary}
\usage{
endPointToxCreate(pCodeInfoDF, casrn_pCode = "casrn",
  units_pCode = "parameter_units", mlWt_pCode = "mlWt")
}
\arguments{
\item{pCodeInfoDF}{data frame}

\item{casrn_pCode}{character name of casrn column in pCodeInfo}

\item{wData}{data frame with a date column (can be NA)
whose name is defined by the date argument, station column whose name is
defined by the station argument, and the remaining columns are measurements that
contain a chemKey (such as parameter code)}

\item{endPoint}{data frame}

\item{date}{character name of date column in wData}

\item{station}{character name of station column in wData}

\item{class_pCode}{character name of class column in pCodeInfo}

\item{code_pCode}{character name of code column in pCodeInfo}

\item{casrn_ep}{character name of class column in endPoint}
}
\value{
chemicalSummary data frame
}
\description{
Create chemical summary
}
\examples{
pCodeInfo <- pCodeInfo
endPoint <- endPointToxCreate(pCodeInfo)
}

