% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chemical_summary.R
\name{exclude_points}
\alias{exclude_points}
\title{Exclude endPoint/Chem combos}
\usage{
exclude_points(chemicalSummary, exclusion)
}
\arguments{
\item{chemicalSummary}{data frame from \code{graph_chem_data}}

\item{exclusion}{data frame with columns "CAS" and "endPoint"}
}
\description{
Using a dataframe "exclusion", filter out all chemical/endpoint combos
}
\examples{
# This is the example workflow:
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"

full_path <- file.path(path_to_tox, file_name)

tox_list <- create_toxEval(full_path)
\dontrun{
ACClong <- get_ACC(tox_list$chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)
chemicalSummary <- get_chemical_summary(tox_list, ACClong, filtered_ep)
}
# The example workflow takes a bit of time to load and compute, 
# so an example chemicalSummary is included pre-calculated in the package. 
chemicalSummary <- ex_chemSum #loading example data
chemicalSummary <- exclude_points(chemicalSummary, exclusion)
}
