% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/explore_endpoints.R
\name{explore_endpoints}
\alias{explore_endpoints}
\title{Explore endpoint groupings}
\usage{
explore_endpoints(browse = TRUE)
}
\arguments{
\item{browse}{use browser for map rendering}
}
\description{
Open an interactive app
}

