% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_endpoints.R
\name{explore_endpoints}
\alias{explore_endpoints}
\title{Explore data in the Shiny Application}
\usage{
explore_endpoints(browse = TRUE)
}
\arguments{
\item{browse}{Logical. Use browser for running Shiny app.}
}
\description{
Open an interactive app in a browser. Using this
function is a quick and convenient way
to explore data. For more customization, the R-code to
produce each graph and table is displayed in the app. That is
a good starting-point for a custom analysis.
}
