% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_endPoint_info.R
\name{filter_groups}
\alias{filter_groups}
\title{Filter endPoints based on groups and assays.}
\usage{
filter_groups(ep, groupCol = "intended_target_family", assays = c("ATG",
  "NVS", "OT", "TOX21", "CEETOX", "APR", "CLD", "TANGUAY", "NHEERL_PADILLA",
  "NCCT_SIMMONS", "ACEA"), remove_groups = c("Background Measurement",
  "Undefined"))
}
\arguments{
\item{ep}{data frame Endpoint information from ToxCast}

\item{groupCol}{character name of column to use as a group catetory}

\item{assays}{vector of assays to use. Possible values are "ATG","NVS","OT","TOX21","CEETOX", "APR", "BSK",
"CLD","TANGUAY","NHEERL_PADILLA","NCCT_SIMMONS","ACEA". By default, the 
"BSK" (BioSeek) assay is removed.}

\item{remove_groups}{vector of groups to remove}
}
\description{
This function takes the data frame from \code{\link{endPointInfo}} 
and filters the endpoints in 3 steps. First, the user specifies
the "groupCol" which is a column header from \code{\link{endPointInfo}}. 
The default category is intended_target_family. Second, the user specifies the assays to use. By default, the BioSeek
set of assays are removed. Finally, the user can also choose to remove
specific group from the category. The default is to remove "Background Measurement"
and "Undefined", but it is a good idea to check if other groups may 
not be relevant to the study.
}
\examples{
endPointInfo <- endPointInfo
cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)
}
