% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chemical_summary.R
\name{get_chemical_summary}
\alias{get_chemical_summary}
\title{get_chemical_summary}
\usage{
get_chemical_summary(ACClong, filtered_ep, chem.data, chem.site, chem.info,
  exclusion = NULL)
}
\arguments{
\item{ACClong}{data frame with at least columns: CAS, chnm, endPoint, ACC_value}

\item{filtered_ep}{data frame with colums: endPoints, groupCol}

\item{chem.data}{data frame with (at least) columns: CAS, SiteID, Value}

\item{chem.site}{data frame with (at least) columns: SiteID, Short Name}

\item{chem.info}{data frame with (at least) columns: CAS, class}

\item{exclusion}{data frame with columns "CAS" and "endPoint"}
}
\description{
Get ACC values for vector of CAS's
}
\examples{
library(readxl)
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"
full_path <- file.path(path_to_tox, file_name)

chem_data <- read_excel(full_path, sheet = "Data")
chem_info <- read_excel(full_path, sheet = "Chemicals") 
chem_site <- read_excel(full_path, sheet = "Sites")

ACClong <- get_ACC(chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)

chemicalSummary <- get_chemical_summary(ACClong,
                                        filtered_ep,
                                       chem_data, 
                                        chem_site, 
                                        chem_info)
                                        
exclusion <- read_excel(full_path, sheet = "Exclude")
chemicalSummary1 <- get_chemical_summary(ACClong,
                                        filtered_ep,
                                       chem_data, 
                                        chem_site, 
                                        chem_info,
                                        exclusion)
}
