% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chemical_summary.R
\name{get_chemical_summary}
\alias{get_chemical_summary}
\title{get_chemical_summary}
\usage{
get_chemical_summary(ACClong, filtered_ep, tox_list = NULL,
  chem.data = NULL, chem.site = NULL, chem.info = NULL,
  exclusion = NULL)
}
\arguments{
\item{ACClong}{data frame with at least columns: CAS, chnm, endPoint, ACC_value}

\item{filtered_ep}{data frame with colums: endPoints, groupCol}

\item{tox_list}{list with data frames for chem_data, chem_info, chem_site, 
and optionally exclusions and benchmarks. Created with \code{\link{create_toxEval}}}

\item{chem.data}{data frame with (at least) columns: CAS, SiteID, Value}

\item{chem.site}{data frame with (at least) columns: SiteID, Short Name}

\item{chem.info}{data frame with (at least) columns: CAS, class}

\item{exclusion}{OPTIONAL data frame with (at least) columns: CAS and endPoint}
}
\description{
Get ACC values for vector of CAS's
}
\examples{
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"
full_path <- file.path(path_to_tox, file_name)

tox_list <- create_toxEval(full_path)

ACClong <- get_ACC(tox_list$chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)

chemicalSummary <- get_chemical_summary(ACClong, filtered_ep,
                                        tox_list)
                                        
library(readxl)
chem_data <- read_excel(full_path, sheet = "Data")
chem_info <- read_excel(full_path, sheet = "Chemicals") 
chem_site <- read_excel(full_path, sheet = "Sites")
                                         
chemicalSummary1 <- get_chemical_summary(ACClong, filtered_ep,
                                         chem.data = chem_data, 
                                         chem.site = chem_site, 
                                        chem.info = chem_info)
}
