% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chemical_summary.R
\name{get_chemical_summary}
\alias{get_chemical_summary}
\title{get_chemical_summary}
\usage{
get_chemical_summary(ACClong, filtered_ep, chem.data, chem.site, chem.info)
}
\arguments{
\item{ACClong}{data frame with columns: casn, chnm, MlWt, endPoint, ACC_value}

\item{filtered_ep}{vector of end points}

\item{chem.data}{data frame with (at least) columns: CAS, SiteID, Value}

\item{chem.site}{data frame with (at least) columns: SiteID, Short Name}

\item{chem.info}{data frame with (at least) columns: CAS, class}
}
\description{
Get ACC values for vector of CAS's
}
\examples{
CAS <- c("121-00-6","136-85-6","80-05-7","84-65-1","5436-43-1","126-73-8")
ACClong <- get_ACC(CAS)
ACClong <- remove_flags(ACClong)
cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)

}
