% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group_boxplots.R
\name{graphData}
\alias{graphData}
\title{graphData}
\usage{
graphData(chemicalSummary, category = "Biological", manual_remove = NULL)
}
\arguments{
\item{chemicalSummary}{data frame}
}
\description{
Just another fancy ggplot2 axis labeler.
}
\keyword{internal}
