% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hits_by_groupings.R
\name{hits_by_groupings_DT}
\alias{hits_by_groupings_DT}
\alias{hits_by_groupings}
\title{Biological hits per category}
\usage{
hits_by_groupings_DT(chemicalSummary, category = "Biological",
  mean_logic = FALSE, hit_threshold = 0.1)

hits_by_groupings(chemicalSummary, category, mean_logic = FALSE,
  hit_threshold = 0.1)
}
\arguments{
\item{chemicalSummary}{data frame from \code{get_chemical_summary}}

\item{category}{either "Biological", "Chemical Class", or "Chemical"}

\item{mean_logic}{logical \code{TRUE} is mean, \code{FALSE} is maximum}

\item{hit_threshold}{numeric threshold defining a "hit"}
}
\description{
These functions create a table with one row per category("Biological", 
"Chemical", or "Chemical Class"). The columns are always the "Biological" 
groupings. The values in the table are how many sites have hits (based on 
hit_threshold), for that particular "Biological"/category combination. Note, 
if the user chooses "Biological", it is a simple 2-column table of 
"Biological" groupings and number of sites (nSites).
}
\details{
The tables show slightly different results for a single site, showing the 
number of samples with hits (instead of number of sites).
}
\examples{
# This is the example workflow:
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"

full_path <- file.path(path_to_tox, file_name)

tox_list <- create_toxEval(full_path)
\dontrun{
ACClong <- get_ACC(tox_list$chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)
chemicalSummary <- get_chemical_summary(tox_list, ACClong, filtered_ep)

site_df <- hits_by_groupings(chemicalSummary, category = "Biological")
hits_by_groupings_DT(chemicalSummary, category = "Biological")
hits_by_groupings_DT(chemicalSummary, category = "Chemical Class")
hits_by_groupings_DT(chemicalSummary, category = "Chemical")
}
}
