% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_tox_sum.R
\name{hits_summary_DT}
\alias{hits_summary_DT}
\alias{hits_summary}
\title{hits_summary_DT}
\usage{
hits_summary_DT(chemicalSummary, category = "Biological",
  mean_logic = FALSE, hit_threshold = 0.1)

hits_summary(chemicalSummary, category, hit_threshold = 0.1)
}
\arguments{
\item{chemicalSummary}{data frame from \code{get_chemical_summary}}

\item{category}{either "Biological", "Chemical Class", or "Chemical"}

\item{mean_logic}{logical \code{TRUE} is mean, \code{FALSE} is maximum}

\item{hit_threshold}{numeric threshold defining a "hit"}
}
\value{
data frame with columns "Hits per Sample", "Individual Hits",
"nSample", "site", and "category"
}
\description{
Table of sums
}
\examples{
# This is the example workflow:
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"

full_path <- file.path(path_to_tox, file_name)

tox_list <- create_toxEval(full_path)
\dontrun{
ACClong <- get_ACC(tox_list$chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)
chemicalSummary <- get_chemical_summary(tox_list, ACClong, filtered_ep)

stats_group <- hits_summary(chemicalSummary, "Biological")

hits_summary_DT(chemicalSummary, category = "Biological")
hits_summary_DT(chemicalSummary, category = "Chemical Class")
hits_summary_DT(chemicalSummary, category = "Chemical")
}
}
