% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMap.R
\name{make_tox_map}
\alias{make_tox_map}
\alias{map_tox_data}
\title{Create an interactive map of the data}
\usage{
make_tox_map(chemicalSummary, chem_site, category = "Biological",
  mean_logic = FALSE, sum_logic = TRUE)

map_tox_data(chemicalSummary, chem_site, category = "Biological",
  mean_logic = FALSE, sum_logic = TRUE)
}
\arguments{
\item{chemicalSummary}{data frame from \code{get_chemical_summary}}

\item{chem_site}{data frame with at least columns SiteID, site_grouping, 
Short Name, dec_lon, and dec_lat}

\item{category}{either "Biological", "Chemical Class", or "Chemical"}

\item{mean_logic}{logical.  TRUE takes the mean sample of each site,
FALSE takes the maximum sample of each site.}

\item{sum_logic}{logical. TRUE sums the EARs in a specified grouping,
FALSE does not. FALSE may be better for traditional benchmarks as
opposed to ToxCast benchmarks.}
}
\description{
This function takes the data and summarizes it based on the 
selected 'category' ("Chemical Class", "Chemical", or "Biological") using
the \code{map_tox_data} function. Then the summarized data is plotted on
an interactive "leaflet" map. This is the only function that requires
"dec_lon" and "dec_lat" (decimal longitude and decimal latitude) in the 
chem_site argument.
}
\examples{
# This is the example workflow:
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"

full_path <- file.path(path_to_tox, file_name)
tox_list <- create_toxEval(full_path)

\dontrun{

ACClong <- get_ACC(tox_list$chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)
chemicalSummary <- get_chemical_summary(tox_list, ACClong, filtered_ep)

make_tox_map(chemicalSummary, tox_list$chem_site, "Biological")   
make_tox_map(chemicalSummary, tox_list$chem_site, "Chemical Class")
make_tox_map(chemicalSummary, tox_list$chem_site, "Chemical") 
}
}
