% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMap.R
\name{make_tox_map}
\alias{make_tox_map}
\alias{map_tox_data}
\title{make_tox_map}
\usage{
make_tox_map(chemicalSummary, chem_site, category = "Biological",
  mean_logic = FALSE)

map_tox_data(chemicalSummary, chem_site, category = "Biological",
  mean_logic = FALSE)
}
\arguments{
\item{chemicalSummary}{data frame from \code{get_chemical_summary}}

\item{chem_site}{data frame with at least columns SiteID, site_grouping, and Short Name}

\item{category}{either "Biological", "Chemical Class", or "Chemical"}

\item{mean_logic}{logical \code{TRUE} is mean, \code{FALSE} is maximum}
}
\description{
make_tox_map
}
\examples{
# This is the example workflow:
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"

full_path <- file.path(path_to_tox, file_name)
tox_list <- create_toxEval(full_path)

\dontrun{

ACClong <- get_ACC(tox_list$chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)
chemicalSummary <- get_chemical_summary(tox_list, ACClong, filtered_ep)

make_tox_map(chemicalSummary, tox_list$chem_site, "Biological")   
make_tox_map(chemicalSummary, tox_list$chem_site, "Chemical Class")
make_tox_map(chemicalSummary, tox_list$chem_site, "Chemical") 
}
}
