% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chemical_boxplots.R
\name{plot_chemical_boxplots}
\alias{plot_chemical_boxplots}
\title{plot_chemical_boxplots}
\usage{
plot_chemical_boxplots(graphData)
}
\arguments{
\item{graphData}{data frame from \code{graph_chem_data}}
}
\description{
Plot boxplot of chemicals
}
\examples{
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"
full_path <- file.path(path_to_tox, file_name)

chem_data <- read_excel(full_path, sheet = "Data")
chem_info <- read_excel(full_path, sheet = "Chemicals") 
chem_site <- read_excel(full_path, sheet = "Sites")
ACClong <- get_ACC(chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)

chemicalSummary <- get_chemical_summary(ACClong,
                                        filtered_ep,
                                       chem_data, 
                                        chem_site, 
                                        chem_info)

graphData <- graph_chem_data(chemicalSummary)
plot_chemical_boxplots(graphData)
}
