% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chemical_boxplots.R
\name{plot_chemical_boxplots}
\alias{plot_chemical_boxplots}
\title{plot_chemical_boxplots}
\usage{
plot_chemical_boxplots(chemicalSummary, manual_remove = NULL,
  mean_logic = FALSE, plot_ND = TRUE)
}
\arguments{
\item{chemicalSummary}{data frame from \code{graph_chem_data}}

\item{manual_remove}{vector of categories to remove}

\item{mean_logic}{logical \code{TRUE} is mean, \code{FALSE} is maximum}

\item{plot_ND}{logical whether or not to plot the non-detects}
}
\description{
Plot boxplot of chemicals
}
\examples{
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"
full_path <- file.path(path_to_tox, file_name)

tox_list <- create_toxEval(full_path)

ACClong <- get_ACC(tox_list$chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)

chemicalSummary <- get_chemical_summary(ACClong, filtered_ep,
                                        tox_list)
                                        
plot_chemical_boxplots(chemicalSummary)
}
