% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heat_chemical.R
\name{plot_heat_chemicals}
\alias{plot_heat_chemicals}
\title{plot_heat_chemicals}
\usage{
plot_heat_chemicals(graphData, chem_site)
}
\arguments{
\item{graphData}{data frame from \code{graph_chem_data}}

\item{chem_site}{data frame with at least columns SiteID, site_grouping,  and Short Name}
}
\description{
Plot heat map
}
\examples{
# This is the example workflow:
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"

full_path <- file.path(path_to_tox, file_name)

tox_list <- create_toxEval(full_path)
\dontrun{
ACClong <- get_ACC(tox_list$chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)
chemicalSummary <- get_chemical_summary(tox_list, ACClong, filtered_ep)
}
# The example workflow takes a bit of time to load and compute, 
# so an example chemicalSummary is included pre-calculated in the package. 

chemicalSummary <- ex_chemSum #loading example data

graphData <- graph_chem_data(chemicalSummary)
plot_heat_chemicals(graphData, tox_list$chem_site)

#Order the site_groupings:
tox_list$chem_site$site_grouping <- factor(tox_list$chem_site$site_grouping,
              levels=c("Lake Superior",
              "Lake Michigan",
              "Lake Huron",
              "Lake Erie",
              "Lake Ontario"))
              
plot_heat_chemicals(graphData, tox_list$chem_site)

#Order sites:
sitesOrdered <- c("StLouis","Nemadji","WhiteWI","Bad","Montreal",
"PresqueIsle","Ontonagon","Sturgeon","Tahquamenon","Burns",
"IndianaHC","StJoseph","PawPaw","Kalamazoo","GrandMI",
"Milwaukee","Muskegon","WhiteMI","PereMarquette","Manitowoc",
"Manistee","Fox","Oconto","Peshtigo","Menominee",
"Indian","Cheboygan","Ford","Escanaba","Manistique",
"ThunderBay","AuSable","Rifle","Saginaw","BlackMI",
"Clinton","Rouge","HuronMI","Raisin","Maumee",
"Portage","Sandusky","HuronOH","Vermilion","BlackOH",
"Rocky","Cuyahoga","GrandOH","Cattaraugus","Tonawanda",
"Genesee","Oswego","BlackNY","Oswegatchie","Grass",
"Raquette","StRegis")

tox_list$chem_site$`Short Name` <- factor(tox_list$chem_site$`Short Name`,
              levels = sitesOrdered)
              
plot_heat_chemicals(graphData, tox_list$chem_site)
}
