% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group_boxplots.R
\name{plot_tox_boxplots}
\alias{plot_tox_boxplots}
\title{plot_tox_boxplots}
\usage{
plot_tox_boxplots(chemicalSummary, filtered_ep, category = "Biological",
  manual_remove = NULL)
}
\arguments{
\item{chemicalSummary}{data frame from \code{get_chemical_summary}}

\item{filtered_ep}{data frame from \code{filter_groups}}

\item{category}{either "Biological", "Chemical Class", or "Chemical"}

\item{manual_remove}{vector of categories to remove}
}
\description{
Plot boxplot of groups
}
\examples{
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"
full_path <- file.path(path_to_tox, file_name)

chem_data <- read_excel(full_path, sheet = "Data")
chem_info <- read_excel(full_path, sheet = "Chemicals") 
chem_site <- read_excel(full_path, sheet = "Sites")
ACClong <- get_ACC(chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)

chemicalSummary <- get_chemical_summary(ACClong,
                                        filtered_ep,
                                       chem_data, 
                                        chem_site, 
                                        chem_info)
plot_tox_boxplots(chemicalSummary, filtered_ep, "Biological")   
plot_tox_boxplots(chemicalSummary, filtered_ep, "Chemical Class")
plot_tox_boxplots(chemicalSummary, filtered_ep, "Chemical") 
}
