% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chemical_boxplots.R,
%   R/plot_group_boxplots.R
\name{graph_chem_data}
\alias{graph_chem_data}
\alias{plot_tox_boxplots}
\alias{tox_boxplot_data}
\title{Grouped Boxplots}
\usage{
graph_chem_data(chemicalSummary, manual_remove = NULL, mean_logic = FALSE)

plot_tox_boxplots(chemicalSummary, category = "Biological",
  manual_remove = NULL, mean_logic = FALSE, plot_ND = TRUE,
  font_size = NA, title = NA, pallette = NA, hit_threshold = NA)

tox_boxplot_data(chemicalSummary, category = "Biological",
  manual_remove = NULL, mean_logic = FALSE)
}
\arguments{
\item{chemicalSummary}{data frame from \code{get_chemical_summary}}

\item{manual_remove}{vector of categories to remove}

\item{mean_logic}{character. Options are "mean", "max", or "noSum". 
TRUE will default to "mean" and FALSE to "max". The default value is "mean". 
The most appropriate use of "noSum" is for non-ToxCast benchmarks. In this case
the values plotted are the overall max of the sample (not the max of the sum
of the sample).}

\item{category}{either "Biological", "Chemical Class", or "Chemical"}

\item{plot_ND}{logical whether or not to plot the non-detects}

\item{font_size}{numeric to adjust the axis font size}

\item{title}{character title for plot.}

\item{pallette}{vector of color pallette for fill. Can be a named vector
to specify specific color for specific categories.}

\item{hit_threshold}{numeric threshold defining a "hit"}
}
\description{
This function creates a set of boxplots based on the original input data 
modified by the processing steps above, and the choice of several input options. 
The result includes boxplots of the maximum or mean value per site based on 
the mean_logic argument This ensures that each site is represented equally 
regardless of how many samples are available per site. Boxplots are generated 
by the chosen category. Categories include "Biological", "Chemical Class", or 
"Chemical". Biological refers to the chosen toxCast annotation as defined 
in the groupCol argument of the \code{filter_groups} function. Chemical Class 
refers to the groupings of chemicals as defined in the "Class" column of the 
"Chemicals" tab of the input file. Choosing Chemical Class will generate a 
separate boxplot for each unique class. Choosing Chemical will generate a 
separate boxplot for each individual chemical in the data set.
}
\details{
The difference in the single-site output is that instead of listing the number 
of sites, it lists the number of unique chemical/endpoint combinations used to 
create the box plot.
}
\examples{
# This is the example workflow:
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"

full_path <- file.path(path_to_tox, file_name)

\dontrun{
tox_list <- create_toxEval(full_path)
ACClong <- get_ACC(tox_list$chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)
chemicalSummary <- get_chemical_summary(tox_list, ACClong, filtered_ep)
plot_tox_boxplots(chemicalSummary, "Biological")   
plot_tox_boxplots(chemicalSummary, "Chemical Class")
plot_tox_boxplots(chemicalSummary, "Chemical")

 # To turn off clipping:
class_plot <- plot_tox_boxplots(chemicalSummary, "Chemical Class")
gb <- ggplot2::ggplot_build(class_plot)
gt <- ggplot2::ggplot_gtable(gb)

gt$layout$clip[gt$layout$name=="panel"] <- "off"

grid::grid.draw(gt) 

cbPalette <- c("#999999", "#E69F00", "#56B4E9", "#009E73", "#F0E442", 
               "#0072B2", "#D55E00", "#CC79A7")
graphData <- tox_boxplot_data(chemicalSummary = chemicalSummary,
                              category = "Biological") 
cbValues <- colorRampPalette(cbPalette)(length(levels(graphData$category)))
names(cbValues) <- levels(graphData$category)

plot_tox_boxplots(chemicalSummary, 
                  hit_threshold = 0.1,
                  category = "Biological",
                  pallette = cbValues,
                  title = 'Maximum EAR per site, grouped by biological activity groupings') 


}
}
