% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tox_endpoints.R
\name{plot_tox_endpoints}
\alias{plot_tox_endpoints}
\title{plot_tox_boxplots}
\usage{
plot_tox_endpoints(chemicalSummary, category = "Biological",
  filterBy = "All", manual_remove = NULL, hit_threshold = 0.1,
  mean_logic = FALSE)
}
\arguments{
\item{chemicalSummary}{data frame from \code{get_chemical_summary}}

\item{category}{either "Biological", "Chemical Class", or "Chemical"}

\item{filterBy}{character either "All" or one of the filtered categories.}

\item{manual_remove}{vector of categories to remove}

\item{hit_threshold}{numeric threshold defining a "hit"}

\item{mean_logic}{logical \code{TRUE} is mean, \code{FALSE} is maximum}
}
\description{
Plot boxplot of groups
}
\examples{
library(readxl)
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"
full_path <- file.path(path_to_tox, file_name)

chem_data <- read_excel(full_path, sheet = "Data")
chem_info <- read_excel(full_path, sheet = "Chemicals") 
chem_site <- read_excel(full_path, sheet = "Sites")
ACClong <- get_ACC(chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)

chemicalSummary <- get_chemical_summary(ACClong,
                                        filtered_ep,
                                       chem_data, 
                                        chem_site, 
                                        chem_info)
 plot_tox_endpoints(chemicalSummary, filterBy = "Cell Cycle")
 plot_tox_endpoints(chemicalSummary, category = "Chemical Class", filterBy = "PAH")
 plot_tox_endpoints(chemicalSummary, category = "Chemical", filterBy = "Atrazine")
}
