% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heat_chemical.R
\name{plot_tox_heatmap}
\alias{plot_tox_heatmap}
\title{plot_tox_heatmap}
\usage{
plot_tox_heatmap(chemicalSummary, chem_site, category = "Biological",
  breaks = c(1e-05, 1e-04, 0.001, 0.01, 0.1, 1, 5), manual_remove = NULL,
  mean_logic = FALSE, plot_ND = TRUE, font_size = NA, title = NA)
}
\arguments{
\item{chemicalSummary}{data frame from \code{get_chemical_summary}}

\item{chem_site}{data frame with at least columns SiteID, site_grouping,  and Short Name}

\item{category}{either "Biological", "Chemical Class", or "Chemical"}

\item{manual_remove}{vector of categories to remove}

\item{mean_logic}{logical \code{TRUE} is mean, \code{FALSE} is maximum}

\item{plot_ND}{logical whether or not to plot the non-detects}

\item{font_size}{numeric to adjust the axis font size}

\item{title}{character title for plot.}
}
\description{
Plot heat map
}
\examples{
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"
full_path <- file.path(path_to_tox, file_name)
\dontrun{
tox_list <- create_toxEval(full_path)

ACClong <- get_ACC(tox_list$chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)

chemicalSummary <- get_chemical_summary(tox_list, ACClong, filtered_ep)
                                        
#Order the site_groupings:
tox_list$chem_site$site_grouping <- factor(tox_list$chem_site$site_grouping,
              levels=c("Lake Superior",
              "Lake Michigan",
              "Lake Huron",
              "Lake Erie",
              "Lake Ontario"))

#Order sites:
sitesOrdered <- c("StLouis","Nemadji","WhiteWI","Bad","Montreal",
"PresqueIsle","Ontonagon","Sturgeon","Tahquamenon","Burns",
"IndianaHC","StJoseph","PawPaw","Kalamazoo","GrandMI",
"Milwaukee","Muskegon","WhiteMI","PereMarquette","Manitowoc",
"Manistee","Fox","Oconto","Peshtigo","Menominee",
"Indian","Cheboygan","Ford","Escanaba","Manistique",
"ThunderBay","AuSable","Rifle","Saginaw","BlackMI",
"Clinton","Rouge","HuronMI","Raisin","Maumee",
"Portage","Sandusky","HuronOH","Vermilion","BlackOH",
"Rocky","Cuyahoga","GrandOH","Cattaraugus","Tonawanda",
"Genesee","Oswego","BlackNY","Oswegatchie","Grass",
"Raquette","StRegis")

tox_list$chem_site$`Short Name` <- factor(tox_list$chem_site$`Short Name`,
              levels = sitesOrdered)
              
plot_tox_heatmap(chemicalSummary, 
                 tox_list$chem_site, 
                 category = "Biological",
                 manual_remove = "Undefined")
plot_tox_heatmap(chemicalSummary, tox_list$chem_site, category = "Chemical Class")
plot_tox_heatmap(chemicalSummary, tox_list$chem_site, category = "Chemical")
}
}
