% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tox_stacks.R
\name{plot_tox_stacks}
\alias{plot_tox_stacks}
\title{plot_tox_boxplots}
\usage{
plot_tox_stacks(chemicalSummary, chem_site, category = "Biological",
  mean_logic = FALSE, manual_remove = NULL, include_legend = TRUE)
}
\arguments{
\item{chemicalSummary}{data frame from \code{get_chemical_summary}}

\item{chem_site}{data frame with at least columns SiteID, site_grouping,  and Short Name}

\item{category}{either "Biological", "Chemical Class", or "Chemical"}

\item{mean_logic}{logical \code{TRUE} is mean, \code{FALSE} is maximum}

\item{manual_remove}{vector of categories to remove}

\item{include_legend}{logical to include legend or not}
}
\description{
Plot boxplot of groups
}
\examples{
# This is the example workflow:
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"

full_path <- file.path(path_to_tox, file_name)

tox_list <- create_toxEval(full_path)
\dontrun{
ACClong <- get_ACC(tox_list$chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)
chemicalSummary <- get_chemical_summary(tox_list, ACClong, filtered_ep)
}
# The example workflow takes a bit of time to load and compute, 
# so an example chemicalSummary is included pre-calculated in the package. 
chemicalSummary <- ex_chemSum #loading example data
                                        
plot_tox_stacks(chemicalSummary, tox_list$chem_site, "Biological")   
plot_tox_stacks(chemicalSummary, tox_list$chem_site, "Chemical Class")
plot_tox_stacks(chemicalSummary, tox_list$chem_site, "Chemical", include_legend = FALSE) 
}
