% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tox_stacks.R
\name{plot_tox_stacks}
\alias{plot_tox_stacks}
\title{plot_tox_boxplots}
\usage{
plot_tox_stacks(chemicalSummary, chem_site, category = "Biological",
  mean_logic = FALSE, manual_remove = NULL, include_legend = TRUE)
}
\arguments{
\item{chemicalSummary}{data frame from \code{get_chemical_summary}}

\item{chem_site}{data frame with at least columns SiteID, site_grouping,  and Short Name}

\item{category}{either "Biological", "Chemical Class", or "Chemical"}

\item{mean_logic}{logical \code{TRUE} is mean, \code{FALSE} is maximum}

\item{manual_remove}{vector of categories to remove}
}
\description{
Plot boxplot of groups
}
\examples{
library(readxl)
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"
full_path <- file.path(path_to_tox, file_name)

chem_data <- read_excel(full_path, sheet = "Data")
chem_info <- read_excel(full_path, sheet = "Chemicals") 
chem_site <- read_excel(full_path, sheet = "Sites")
ACClong <- get_ACC(chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)

chemicalSummary <- get_chemical_summary(ACClong,
                                        filtered_ep,
                                       chem_data, 
                                        chem_site, 
                                        chem_info)
plot_tox_stacks(chemicalSummary, chem_site, "Biological")   
plot_tox_stacks(chemicalSummary, chem_site, "Chemical Class")
plot_tox_stacks(chemicalSummary, chem_site, "Chemical", include_legend = FALSE) 
}
