% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siteSumm.R
\name{siteSumm}
\alias{siteSumm}
\title{Create site summary}
\usage{
siteSumm(chemicalSummary, newSiteKey)
}
\arguments{
\item{wData}{data frame with a date column (can be NA)
whose name is defined by the date argument, station column whose name is
defined by the station argument, and the remaining columns are measurements that
contain a chemKey (such as parameter code)}

\item{pCodeInfoDF}{data frame}

\item{endPoint}{data frame}

\item{date}{character name of date column in wData}

\item{station}{character name of station column in wData}

\item{casrn_pCode}{character name of casrn column in pCodeInfo}

\item{class_pCode}{character name of class column in pCodeInfo}

\item{code_pCode}{character name of code column in pCodeInfo}

\item{casrn_ep}{character name of class column in endPoint}
}
\description{
Create site summary
}
\examples{
wData <- wData
pCodeInfo <- pCodeInfo
packagePath <- system.file("extdata", package="toxEval")
filePath <- file.path(packagePath, "stationINFO.RData")
load(file=filePath)
newSiteKey <- setNames(stationINFO$shortName, stationINFO$fullSiteID)
endPoint <- endPointToxCreate(pCodeInfo)
chemicalSummary <- chemSummBasic(wData,pCodeInfo,endPoint)
siteSummary <- siteSumm(chemicalSummary,newSiteKey)
}

