% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_tox_rank.R
\name{statsOfColumns}
\alias{statsOfColumns}
\title{statsOfColumns}
\usage{
statsOfColumns(chemicalSummary, category, hit_threshold = 0.1,
  mean_logic = FALSE)
}
\arguments{
\item{chemicalSummary}{data frame}

\item{category}{character}

\item{hit_threshold}{numeric}

\item{mean_logic}{logical}
}
\description{
Summarize data for most graphs/tables
}
\examples{
# This is the example workflow:
path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"

full_path <- file.path(path_to_tox, file_name)

tox_list <- create_toxEval(full_path)
\dontrun{
ACClong <- get_ACC(tox_list$chem_info$CAS)
ACClong <- remove_flags(ACClong)

cleaned_ep <- clean_endPoint_info(endPointInfo)
filtered_ep <- filter_groups(cleaned_ep)
chemicalSummary <- get_chemical_summary(tox_list, ACClong, filtered_ep)

stats_df <- statsOfColumns(chemicalSummary, "Biological")
} 
}
