% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxEval.R
\docType{data}
\name{tox_chemicals}
\alias{tox_chemicals}
\title{ToxCast Chemical Information}
\value{
data frame with columns:    
"Substance_Name","Substance_CASRN",
"Structure_MolWt"
}
\description{
Downloaded on October 2015 from ToxCast. The file name of the
raw data was "TOX21IDs_v4b_23Oct2014_QCdetails.xlsx", 
from the US EPA DSSTox DATA RELEASE OCTOBER 2015. At the time
of toxEval package release, this information was found:
\url{https://www.epa.gov/chemical-research/exploring-toxcast-data-downloadable-data}
in the section marked "Download ToxCast Chemical Information". This 
was in the "ToxCast & Tox21 Chemicals Distributed Structure-Searchable Toxicity Database (DSSTox files)"
data set.
}
\examples{
head(tox_chemicals)
}
\keyword{datasets}
