% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxEval.R
\docType{data}
\name{ToxCast_ACC}
\alias{ToxCast_ACC}
\title{ACC values included with toxEval. See \code{vignette("Setting up toxEval package data", package = "toxEval")}
for more information on how the data was aggregated.}
\source{
\url{https://www.epa.gov/comptox-tools/exploring-toxcast-data}
}
\value{
data frame with columns CAS, chnm (chemical name), flags, endPoint, and ACC (value).
}
\description{
Downloaded on September 2023 from ToxCast. See also
\url{https://www.frontiersin.org/articles/10.3389/ftox.2023.1275980/full}.
}
\examples{
head(ToxCast_ACC)
}
\references{
U.S. EPA. 2023. ToxCast & Tox21 Summary Files. 
Retrieved from \url{https://www.epa.gov/comptox-tools/exploring-toxcast-data}
on September 2023.
}
\keyword{datasets}
