% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxEval.R
\docType{package}
\name{toxEval-package}
\alias{toxEval-package}
\title{Analyze ToxCast data in relation to measured concentrations.}
\description{
\code{toxEval} includes a set of functions to analyze, visualize, and 
organize measured concentration data as it relates to ToxCast data 
(default) or other user-selected chemical-biological interaction 
benchmark data such as water quality criteria. The intent of 
these analyses is to develop a better understanding of the potential 
biological relevance of environmental chemistry data. Results can 
be used to prioritize which chemicals at which sites may be of 
greatest concern. These methods are meant to be used as a screening 
technique to predict potential for biological influence from chemicals 
that ultimately need to be validated with direct biological assays.
}
\details{
\tabular{ll}{
Package: \tab toxEval\cr
Type: \tab Package\cr
License: \tab Unlimited for this package, dependencies have more restrictive licensing.\cr
Copyright: \tab This software is in the public domain because it contains materials
that originally came from the United States Geological Survey, an agency of
the United States Department of Interior. For more information, see the
official USGS copyright policy at
https://www.usgs.gov/visual-id/credit_usgs.html#copyright\cr
LazyLoad: \tab yes\cr
}
}
\author{
Laura De Cicco \email{ldecicco@usgs.gov}. Steven Corsi
}
\keyword{ToxCast}
