% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxEval.R
\docType{package}
\name{toxEval-package}
\alias{toxEval-package}
\title{Evaluation of ToxCast data with measured concentrations}
\description{
\tabular{ll}{
Package: \tab toxEval\cr
Type: \tab Package\cr
License: \tab Unlimited for this package, dependencies have more restrictive licensing.\cr
Copyright: \tab This software is in the public domain because it contains materials
that originally came from the United States Geological Survey, an agency of
the United States Department of Interior. For more information, see the
official USGS copyright policy at
https://www.usgs.gov/visual-id/credit_usgs.html#copyright\cr
LazyLoad: \tab yes\cr
}
}
\details{
Initial code for studying ToxCast data in relation to measured concentrations.
}
\author{
Steven Corsi \email{srcorsi@usgs.gov}, Laura De Cicco \email{ldecicco@usgs.gov}
}
\keyword{ToxCast}
